revert some if statements

+++ modules/tfe_init/templates/tfe.sh.tpl
@@ -101,11 +101,13 @@ echo "[$(date +"%FT%T")] [Terraform Enterprise] Skipping TlsBootstrapKey configu
 #------------------------------------------------------------------------------
 ca_certificate_directory="/dev/null"
 
-%{ if "$distribution" == "rhel" ~}
-ca_certificate_directory=/usr/share/pki/ca-trust-source/anchors
-%{ else ~}
-ca_certificate_directory=/usr/local/share/ca-certificates/extra
-%{ endif ~}
+if [[ $distribution == "rhel" ]]
+then
+	ca_certificate_directory=/usr/share/pki/ca-trust-source/anchors
+else
+	ca_certificate_directory=/usr/local/share/ca-certificates/extra
+fi
+
 ca_cert_filepath="$ca_certificate_directory/tfe-ca-certificate.crt"
 
 %{ if ca_certificate_secret_id != null ~}
@@ -114,25 +116,30 @@ ca_certificate_data_b64=$(get_base64_secrets ${ca_certificate_secret_id})
 
 mkdir -p $ca_certificate_directory
 echo $ca_certificate_data_b64 | base64 --decode > $ca_cert_filepath
+
 %{ else ~}
 echo "[$(date +"%FT%T")] [Terraform Enterprise] Skipping CA certificate configuration" | tee -a $log_pathname
 %{ endif ~}
 
-%{ if -f "$ca_cert_filepath" && "$distribution" == "rhel" ~}
-update-ca-trust
-%{ elsif -f "$ca_cert_filepath" && "$distribution" == "ubuntu" ~}
-update-ca-certificates
-%{ endif ~}
+if [ -f "$ca_cert_filepath" ]
+then
+	if [[ $distribution == "rhel" ]]
+	then
+		update-ca-trust
+	else
+		update-ca-certificates
+	fi
 
-%{ if -f "$ca_cert_filepath" ~}
-jq ". + { ca_certs: { value: \"$(/bin/cat $ca_cert_filepath)\" } }" -- $tfe_settings_path > $tfe_settings_file.updated
-cp ./$tfe_settings_file.updated $tfe_settings_path
-%{ endif ~}
+	jq ". + { ca_certs: { value: \"$(/bin/cat $ca_cert_filepath)\" } }" -- $tfe_settings_path > $tfe_settings_file.updated
+	cp ./$tfe_settings_file.updated $tfe_settings_path
+fi
 
 # -----------------------------------------------------------------------------
 # Resize RHEL logical volume (if Azure environment)
 # -----------------------------------------------------------------------------
-%{ if cloud == "azurerm" && "$distribution" == "rhel" ~}
+%{ if cloud == "azurerm" ~}
+if [[ $distribution == "rhel" ]]
+then
 echo "[$(date +"%FT%T")] [Terraform Enterprise] Resize RHEL logical volume" | tee -a $log_pathname
 # Because Microsoft is publishing only LVM-partitioned images, it is necessary to partition it to the specs that TFE requires.
 # First, extend the partition to fill available space
@@ -142,6 +149,7 @@ pvresize /dev/disk/azure/root-part4
 # Then resize the logical volumes to meet TFE specs
 lvresize -r -L 10G /dev/mapper/rootvg-rootlv
 lvresize -r -L 40G /dev/mapper/rootvg-varlv
+fi
 %{ endif ~}
 
 # -----------------------------------------------------------------------------
@@ -162,27 +170,28 @@ replicated_directory="/etc/replicated"
 # Bootstrap airgapped environment with prerequisites (for dev/test environments)
 echo "[Terraform Enterprise] Installing Docker Engine from Repository for Bootstrapping an Airgapped Installation" | tee -a $log_pathname
 
-	%{ if "$distribution" == "rhel" ~}
-	yum install --assumeyes yum-utils
-	yum-config-manager --add-repo https://download.docker.com/linux/rhel/docker-ce.repo
-	yum install --assumeyes docker-ce docker-ce-cli containerd.io
-	%{ else ~}
-	apt-get --assume-yes update
-	apt-get --assume-yes install \
-		ca-certificates \
-		curl \
-		gnupg \
-		lsb-release
-	curl --fail --silent --show-error --location https://download.docker.com/linux/ubuntu/gpg \
-		| gpg --dearmor --output /usr/share/keyrings/docker-archive-keyring.gpg
-	echo \
-		"deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] \
-		https://download.docker.com/linux/ubuntu $(lsb_release --codename --short) stable" \
-		| sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-	apt-get --assume-yes update
-	apt-get --assume-yes install docker-ce docker-ce-cli containerd.io
-	apt-get --assume-yes autoremove
-	%{ endif ~}
+if [[ $distribution == "rhel" ]]
+then
+yum install --assumeyes yum-utils
+yum-config-manager --add-repo https://download.docker.com/linux/rhel/docker-ce.repo
+yum install --assumeyes docker-ce docker-ce-cli containerd.io
+elif [[ $distribution == "ubuntu" ]]
+apt-get --assume-yes update
+apt-get --assume-yes install \
+	ca-certificates \
+	curl \
+	gnupg \
+	lsb-release
+curl --fail --silent --show-error --location https://download.docker.com/linux/ubuntu/gpg \
+	| gpg --dearmor --output /usr/share/keyrings/docker-archive-keyring.gpg
+echo \
+	"deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] \
+	https://download.docker.com/linux/ubuntu $(lsb_release --codename --short) stable" \
+	| sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+apt-get --assume-yes update
+apt-get --assume-yes install docker-ce docker-ce-cli containerd.io
+apt-get --assume-yes autoremove
+fi
 
 replicated_filename="replicated.tar.gz"
 replicated_url="https://s3.amazonaws.com/replicated-airgap-work/$replicated_filename"
@@ -233,12 +242,13 @@ $install_pathname \
 # -----------------------------------------------------------------------------
 # Add docker0 to firewalld (for Red Hat instances only)
 # -----------------------------------------------------------------------------
-%{ if "$distribution" == "rhel" ~}
-echo "[$(date +"%FT%T")] [Terraform Enterprise] Disable SELinux (temporary)" | tee -a $log_pathname
-setenforce 0
-echo "[$(date +"%FT%T")] [Terraform Enterprise] Add docker0 to firewalld" | tee -a $log_pathname
-firewall-cmd --permanent --zone=trusted --change-interface=docker0
-firewall-cmd --reload
-echo "[$(date +"%FT%T")] [Terraform Enterprise] Enable SELinux" | tee -a $log_pathname
-setenforce 1
-%{ endif ~}
+if [[ $distribution == "rhel" ]]
+then
+	echo "[$(date +"%FT%T")] [Terraform Enterprise] Disable SELinux (temporary)" | tee -a $log_pathname
+	setenforce 0
+	echo "[$(date +"%FT%T")] [Terraform Enterprise] Add docker0 to firewalld" | tee -a $log_pathname
+	firewall-cmd --permanent --zone=trusted --change-interface=docker0
+	firewall-cmd --reload
+	echo "[$(date +"%FT%T")] [Terraform Enterprise] Enable SELinux" | tee -a $log_pathname
+	setenforce 1
+fi
